+++ src/pactole/combinations/euromillions_combination.py
@@ -111,7 +111,7 @@ class EuroMillionsCombination(LotteryCombination):
             >>> euro_comb.to_dict()
             {'numbers': [3, 15, 22, 28, 44], 'stars': [2, 9]}
         """
-        return {name: component.to_json() for name, component in self._components.items()}
+        return self.dump()
 
     @classmethod
     def from_dict(cls, data: dict) -> EuroMillionsCombination:
@@ -154,7 +154,7 @@ class EuroMillionsCombination(LotteryCombination):
             >>> euro_comb.stars.values
             [2, 9]
         """
-        components = LotteryCombination.from_string(data)
+        components = LotteryCombination.get_components_from_string(data)
         return cls(**components)
 
     @classmethod
@@ -185,5 +185,5 @@ class EuroMillionsCombination(LotteryCombination):
             >>> euro_comb.stars.values
             [2, 9]
         """
-        components = LotteryCombination.from_csv(data)
+        components = LotteryCombination.get_components_from_csv(data)
         return cls(**components)